requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
numpy>=1.21.0
scikit-learn>=1.0.0
faiss-cpu>=1.7.0  # Use faiss-gpu for GPU support
tqdm>=4.65.0
pytest>=7.0.0  # For running tests